feat(anthropic): add HTTPX async adapter support
