+++ src/llm_api_adapter/adapters/anthropic/__init__.py
@@ -1,0 +1,5 @@
+"""Anthropic adapter package."""
+
+from .adapter import AnthropicAdapter
+
+__all__ = ["AnthropicAdapter"]
